extracting interface for mockregistry in scalasdk
